created basic layout for plans page
