+++ src/App.jsx
@@ -5,7 +5,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import SignIn from './pages/SignIn';
 import SignUp from './pages/SignUp';
-import Plans from './pages/Plans';
+import Plans from './pages/plans/Plans';
 
 export default function App() {
   return (
