+++ src/assets/css/global.css
@@ -7,6 +7,7 @@ body {
 
 #root,
 #root > div {
+  display: flex;
   min-height: 100%;
   font-family: 'Roboto', sans-serif;
 }
